Fix broken command availability check
This logic was breaking in CentOS 6.6 box with all dependencies (ovs-vsctl, docker, and uuidgen) installed and working. The check_command_avail function checks to see if the commands exist by running "$command --version". This is flawed logic because it assumes "--version" will work with all of the commands that are listed in the function call. It's much more common (and future proof) to simply check to see that the required binaries are available by running the "which" command. Also, I added a ! to make the control structure a little more terse.

Evidence of the problem with checking --version:
[root@docker ~]# uuidgen --version
uuidgen: invalid option -- '-'
Usage: uuidgen [-r] [-t]
[root@docker ~]# echo $?
1